Create LICENSE
As of May 18th, 2018 the content of the repository is licensed under the terms of the GNU AGPL 3.0 license. The MIT license is included in the MIT-LICENSE file to fulfill the terms of the license of the versions of this repository available prior to May 18th, 2018, and to fulfill the terms of the JQuery license and other included dependencies of this project licensed thereunder. Any changes made on that day, or at any time thereafter shall be licensed under the GNU AGPL 3.0.